Fix docupen camlib compilation warnings.
* cache.c:
  * Add 'size_t' cast to 'ret' in signed-unsigned comparison. From previous lines it is guaranteed that 'ret' is not negative.
* docupen.c:
  * models[] array is unused and therefore commented out.
  * get_file_func(...): fread(...) returns 'size_t' so change 'ret' type.
  * camera_init(...): Remove unused variable.

+++ camlibs/docupen/cache.c
@@ -43,7 +43,7 @@ static bool fill_cache(Camera *camera)
 			break;
 		fwrite(buf, 1, ret, camera->pl->cache);
 		done += ret;
-		if (ret < pktsize)
+		if ((size_t)ret < pktsize)
 			break;
 	}
 	free(buf);
